+++ src/index.ts
@@ -34,6 +34,7 @@ export {
   Schema,
   PathToModel,
   Query,
+  Mutation
 } from './graphql/types';
 export {HasuraSchemaLoader} from './graphql/hasura-schema-loader';
 export {QueryBuilder, mask, batchMutation} from './graphql/query-builder';
