+++ src/types.ts
@@ -82,28 +82,3 @@ export interface Model {
   keySchema: any;
   dataSchema: any;
 }
-
-export interface Mutation {
-  mutation: {
-    [key: string]: {
-      __args: MutationObject;
-      id: boolean;
-    };
-  };
-}
-
-export interface MutationObject {
-  object?: any;
-  data?: any;
-  on_conflict: ConflictClause;
-}
-
-export interface MutationReference {
-  data: any;
-  on_conflict: ConflictClause;
-}
-
-export interface ConflictClause {
-  constraint: EnumType;
-  update_columns: EnumType[];
-}
